create addUsers script

+++ seed/users.ts
@@ -3,7 +3,6 @@ import { faker } from '@faker-js/faker'
 import { User } from '../src/models/User'
 import * as argon2 from 'argon2'
 import { UserRepository } from '../src/repositories/UserRepository'
-import seedConf from '../data/seed.json'
 import { getSchemaSQL } from './utils'
 import path from 'path'
 
@@ -11,34 +10,36 @@ let db: Database
 
 const DAY_IN_MILLIS = 1000*60*60*24
 
-const script = async (): Promise<User[]> => {
+interface SeedUser {
+  username: string,
+  password: string,
+  lockout: boolean,
+}
+
+interface SeedUserConfig {
+  numRandomUsers: number
+  additionalUsers: SeedUser[]
+  shouldClear: boolean  // Clears the existing table if it exists
+}
+
+const script = async (config: SeedUserConfig): Promise<User[]> => {
   console.log('Starting user seeding...')
 
   db = await Database.getInstance()
 
-  console.log('Dropping table...')
-  await db.query(`
-    DROP TABLE IF EXISTS users CASCADE;
-  `)
+  if (config.shouldClear) {
+    console.log('Dropping table...')
+    await db.query(`
+      DROP TABLE IF EXISTS users CASCADE;
+    `)
 
-  console.log('Creating table...')
-  await db.query(await getSchemaSQL(path.resolve(__dirname, './schema/001_users.sql')))
-
+    console.log('Creating table...')
+    await db.query(await getSchemaSQL(path.resolve(__dirname, './schema/001_users.sql')))
+  }
+  
   console.log('Inserting data...')
-  const additionalUsers = [
-    {
-      username: "admin1",
-      password: "strongpassword",
-      lockout: true,
-    },
-    {
-      username: "admin2",
-      password: "weakpassword",
-      lockout: false,
-    }
-  ]
 
-  return Promise.all(Array.from({length: seedConf.NUM_RANDOM_USERS}).map(async () => {
+  return Promise.all(Array.from({length: config.numRandomUsers}).map(async () => {
     const hash = await argon2.hash(faker.internet.password())
 
     return UserRepository.insert({
@@ -46,7 +47,7 @@ const script = async (): Promise<User[]> => {
       passwordHash: hash,
       lockout: faker.datatype.boolean(),
     })
-  }).concat(additionalUsers.map(async (user) => {
+  }).concat(config.additionalUsers.map(async (user) => {
     const hash = await argon2.hash(user.password)
 
     return UserRepository.insert({
